Update install-sdk.md (#559)
As I understand it, pip3 is the Python3 version of pip. On systems with both Python 2 and 3, pip will use Python 2. This affects hosted notebook environments that have both Python 2 and 3 installed, such as Kubeflow notebooks or Google Cloud Notebooks. Since the SDK requires Python 3, I'd recommend changing this to be pip3.